refact: refatoração nas classes java separando as responsabilidades
