+++ backend/src/main/java/br/com/budgetflow/features/planejamentos/service/PlanejamentoService.java
@@ -1,15 +1,5 @@
 package br.com.budgetflow.features.planejamentos.service;
 
-import java.nio.charset.StandardCharsets;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.HexFormat;
-import java.util.List;
-import java.util.Set;
-
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -20,9 +10,6 @@ import br.com.budgetflow.common.exceptions.BusinessRuleException;
 import br.com.budgetflow.common.exceptions.ResourceNotFoundException;
 import br.com.budgetflow.features.categorias.domain.Categoria;
 import br.com.budgetflow.features.categorias.repository.CategoriaRepository;
-import br.com.budgetflow.features.movimentacoes.domain.TransacaoRecorrente;
-import br.com.budgetflow.features.movimentacoes.repository.TransacaoRecorrenteRepository;
-import br.com.budgetflow.features.movimentacoes.service.support.RecorrenciaUtils;
 import br.com.budgetflow.features.periodos.domain.PeriodoFinanceiro;
 import br.com.budgetflow.features.periodos.service.PeriodoFinanceiroService;
 import br.com.budgetflow.features.planejamentos.criteria.PlanejamentoFilterCriteria;
@@ -41,26 +28,29 @@ import br.com.budgetflow.security.SecurityUtils;
 public class PlanejamentoService {
 
     private final PlanejamentoRepository planejamentoRepository;
-    private final TransacaoRecorrenteRepository recorrenteRepository;
     private final CategoriaRepository categoriaRepository;
     private final PeriodoFinanceiroService periodoFinanceiroService;
     private final UserService userService;
     private final PlanejamentoMapper planejamentoMapper;
+    private final PlanejamentoRecorrenteVinculoService recorrenteVinculoService;
+    private final GeracaoPlanejamentosRecorrentesService geracaoPlanejamentosRecorrentesService;
 
     public PlanejamentoService(
             PlanejamentoRepository planejamentoRepository,
-            TransacaoRecorrenteRepository recorrenteRepository,
             CategoriaRepository categoriaRepository,
             PeriodoFinanceiroService periodoFinanceiroService,
             UserService userService,
-            PlanejamentoMapper planejamentoMapper
+            PlanejamentoMapper planejamentoMapper,
+            PlanejamentoRecorrenteVinculoService recorrenteVinculoService,
+            GeracaoPlanejamentosRecorrentesService geracaoPlanejamentosRecorrentesService
     ) {
         this.planejamentoRepository = planejamentoRepository;
-        this.recorrenteRepository = recorrenteRepository;
         this.categoriaRepository = categoriaRepository;
         this.periodoFinanceiroService = periodoFinanceiroService;
         this.userService = userService;
         this.planejamentoMapper = planejamentoMapper;
+        this.recorrenteVinculoService = recorrenteVinculoService;
+        this.geracaoPlanejamentosRecorrentesService = geracaoPlanejamentosRecorrentesService;
     }
 
     @Transactional(readOnly = true)
@@ -68,10 +58,10 @@ public class PlanejamentoService {
         Long userId = SecurityUtils.currentUserId();
         PeriodoFinanceiro periodo = periodoFinanceiroService.resolvePeriodoToTransacao(criteria.getPeriodoId(), userId);
 
-        PlanejamentoFilterCriteria effective = criteria;
-        effective.setPeriodoId(periodo.getId());
+        PlanejamentoFilterCriteria planejamentos = criteria;
+        planejamentos.setPeriodoId(periodo.getId());
 
-        Specification<Planejamento> specification = PlanejamentoSpecification.createSpecification(effective, userId);
+        Specification<Planejamento> specification = PlanejamentoSpecification.createSpecification(planejamentos, userId);
         return planejamentoRepository.findAll(specification, pageable).map(planejamentoMapper::toResponseDTO);
     }
 
@@ -88,7 +78,12 @@ public class PlanejamentoService {
         fillEditableFields(planejamento, request, categoria);
 
         if (request.transacaoRecorrenteId() != null) {
-            applyRecorrenteSyncKey(planejamento, request.transacaoRecorrenteId(), periodo, userId);
+            String chaveSincronizacao = recorrenteVinculoService.resolverChaveSincronizacao(
+                    request.transacaoRecorrenteId(),
+                    periodo,
+                    userId
+            );
+            planejamento.setChaveSincronizacao(chaveSincronizacao);
         }
 
         return planejamentoMapper.toResponseDTO(planejamentoRepository.save(planejamento));
@@ -132,119 +127,7 @@ public class PlanejamentoService {
     public SincronizacaoPlanejamentosResponseDTO sincronizarRecorrentes(Long periodoId) {
         Long userId = SecurityUtils.currentUserId();
         PeriodoFinanceiro periodo = periodoFinanceiroService.resolvePeriodoToTransacao(periodoId, userId);
-        List<Planejamento> novos = new ArrayList<>();
-        Set<String> chavesSincronizadas = new HashSet<>(
-                planejamentoRepository.findChavesSincronizacaoByPeriodoIdAndUserId(periodo.getId(), userId)
-        );
-        int semValor = 0;
-
-        for (TransacaoRecorrente recorrente : recorrenteRepository.findAllByUserId(userId)) {
-            if (recorrente.getValor() == null) {
-                semValor++;
-                continue;
-            }
-            gerarOcorrenciasDoPeriodo(recorrente, periodo, userId, chavesSincronizadas, novos);
-        }
-
-        planejamentoRepository.saveAll(novos);
-        String mensagem = "Sincronização concluída: " + novos.size()
-                + " planejamentos gerados e " + semValor + " recorrências sem valor.";
-        return new SincronizacaoPlanejamentosResponseDTO(novos.size(), semValor, mensagem);
-    }
-
-    private void applyRecorrenteSyncKey(
-            Planejamento planejamento,
-            Long recorrenteId,
-            PeriodoFinanceiro periodo,
-            Long userId
-    ) {
-        TransacaoRecorrente recorrente = recorrenteRepository.findByIdAndUserId(recorrenteId, userId)
-                .orElseThrow(() -> new ResourceNotFoundException("Transação recorrente não encontrada"));
-
-        Set<String> chavesExistentes = planejamentoRepository.findChavesSincronizacaoByPeriodoIdAndUserId(
-                periodo.getId(),
-                userId
-        );
-        LocalDate dataOcorrencia = findFirstAvailableOccurrence(recorrente, periodo, userId, chavesExistentes);
-        if (dataOcorrencia == null) {
-            throw new BusinessRuleException(
-                    "Todas as ocorrências desta recorrência já foram lançadas neste período"
-            );
-        }
-
-        planejamento.setChaveSincronizacao(buildSyncKey(userId, recorrente.getId(), dataOcorrencia));
-    }
-
-    private LocalDate findFirstAvailableOccurrence(
-            TransacaoRecorrente recorrente,
-            PeriodoFinanceiro periodo,
-            Long userId,
-            Set<String> chavesExistentes
-    ) {
-        long indice = 0;
-        while (recorrente.getTotalParcelas() == null || indice < recorrente.getTotalParcelas()) {
-            LocalDate data = RecorrenciaUtils.calcularDataOcorrencia(
-                    recorrente.getDataInicio(),
-                    recorrente.getFrequencia(),
-                    indice
-            );
-
-            if (recorrente.getDataFim() != null && data.isAfter(recorrente.getDataFim())) {
-                break;
-            }
-            if (data.isAfter(periodo.getDataFim())) {
-                break;
-            }
-
-            if (!data.isBefore(periodo.getDataInicio())) {
-                String chave = buildSyncKey(userId, recorrente.getId(), data);
-                if (!chavesExistentes.contains(chave)) {
-                    return data;
-                }
-            }
-            indice++;
-        }
-        return null;
-    }
-
-    private void gerarOcorrenciasDoPeriodo(
-            TransacaoRecorrente recorrente,
-            PeriodoFinanceiro periodo,
-            Long userId,
-            Set<String> chavesSincronizadas,
-            List<Planejamento> novos
-    ) {
-        long indice = 0;
-        while (recorrente.getTotalParcelas() == null || indice < recorrente.getTotalParcelas()) {
-            LocalDate data = RecorrenciaUtils.calcularDataOcorrencia(
-                    recorrente.getDataInicio(),
-                    recorrente.getFrequencia(),
-                    indice
-            );
-
-            if (recorrente.getDataFim() != null && data.isAfter(recorrente.getDataFim())) {
-                break;
-            }
-            if (data.isAfter(periodo.getDataFim())) {
-                break;
-            }
-
-            if (!data.isBefore(periodo.getDataInicio())) {
-                String chave = buildSyncKey(userId, recorrente.getId(), data);
-                if (chavesSincronizadas.add(chave)) {
-                    Planejamento planejamento = new Planejamento();
-                    planejamento.setUser(recorrente.getUser());
-                    planejamento.setPeriodo(periodo);
-                    planejamento.setCategoria(recorrente.getCategoria());
-                    planejamento.setDescricao(recorrente.getDescricao());
-                    planejamento.setValor(recorrente.getValor());
-                    planejamento.setTipoMovimentacao(recorrente.getTipoMovimentacao());
-                    planejamento.setChaveSincronizacao(chave);
-                    novos.add(planejamento);
-                }
-            }
-            indice++;
-        }
+        return geracaoPlanejamentosRecorrentesService.sincronizar(periodo, userId);
     }
 
     private Categoria resolveCategoria(PlanejamentoRequestDTO request, Long userId) {
@@ -271,15 +154,5 @@ public class PlanejamentoService {
     private Planejamento findEntity(Long id, Long userId) {
         return planejamentoRepository.findByIdAndUserIdAndExcluidoFalse(id, userId)
                 .orElseThrow(() -> new ResourceNotFoundException("Planejamento não encontrado"));
-    }
-
-    private String buildSyncKey(Long userId, Long recorrenteId, LocalDate data) {
-        try {
-            MessageDigest digest = MessageDigest.getInstance("SHA-256");
-            String source = userId + ":" + recorrenteId + ":" + data;
-            return HexFormat.of().formatHex(digest.digest(source.getBytes(StandardCharsets.UTF_8)));
-        } catch (NoSuchAlgorithmException ex) {
-            throw new IllegalStateException("SHA-256 indisponível", ex);
-        }
     }
 }
